........
  r14297 | martineau | 2008-11-17 14:00:20 -0500 (Mon, 17 Nov 2008) | 3 lines
  
  BugId: 6240
  * server-src/taper.c: Fix formating.
........


git-svn-id: https://amanda.svn.sourceforge.net/svnroot/amanda/amanda/trunk@1385 a8d146d6-cc15-0410-8900-af154a0219e0

+++ server-src/taper.c
@@ -733,9 +733,9 @@ static gboolean finish_part_attempt(taper_state_t * taper_state,
                 taper_source_predict_parts(dump_info->source),
                 dump_info->level, part_time, (uintmax_t)part_kbytes, part_kbps,
 		consumer_errstr);
-	log_add(L_INFO, "tape %s kb %ld fm %d [OK]\n",
+	log_add(L_INFO, "tape %s kb %lld fm %d [OK]\n",
 		volume_label,
-		((taper_state->total_bytes+(off_t)1023) / (off_t)1024),
+		(long long)((taper_state->total_bytes+(off_t)1023) / (off_t)1024),
 		taper_state->device->file);
 
         /* A problem occured. */
@@ -1211,9 +1211,9 @@ static gboolean process_driver_command(taper_state_t * state) {
     case QUIT:
 	free_cmdargs(cmdargs);
 	if (state->device && state->device->volume_label) {
-	    log_add(L_INFO, "tape %s kb %ld fm %d [OK]\n",
+	    log_add(L_INFO, "tape %s kb %lld fm %d [OK]\n",
 		    state->device->volume_label,
-		    ((state->total_bytes+(off_t)1023) / (off_t)1024),
+		    (long long)((state->total_bytes+(off_t)1023) / (off_t)1024),
 		    state->device->file);
 	}
         return send_quitting(state);
